Reorganize and add unit tests
